+++ client/src/app/(auth)/otp/page.tsx
@@ -1,136 +1,71 @@
 "use client";
-
-import { useCallback, useEffect, useRef, useState } from "react";
-import { useRouter } from "next/navigation";
-
-import axios, { AxiosError } from "axios";
+// import ArrowBackIcon from "@/icons/arrow-back-outline.svg";
 import toast from "react-hot-toast";
 
-const page = () => {
-  const [otp, setOtp] = useState("");
-  const inputRefs = useRef<HTMLInputElement[]>([]);
-  const [isSubmitting, setIsSubmitting] = useState(false);
-  const router = useRouter();
+import OtpInput from "@/app/components/auth-components/otp";
+import Button from "@/app/components/auth-components/Button";
+import { motion } from "framer-motion";
 
-  const submitOTP = useCallback(async () => {
-    if (!otp || isSubmitting) return;
+interface OtpProps {
+  email: string;
+}
 
+const page = ({ email }: OtpProps) => {
+  const handleResubmission = async () => {
     try {
-      setIsSubmitting(true);
-      await axios.post(
-        `${process.env.NEXT_PUBLIC_API_URL}/auth/otp`,
-        {
-          code: otp,
-        },
-        {
-          withCredentials: true,
-        }
-      );
-      router.push("/");
-    } catch (e) {
-      const error = (e as AxiosError).response?.data;
-      const message =
-        (error as { message: string })?.message || "Something went wrong";
-      toast.error(message);
-      setOtp("");
-    } finally {
-      setOtp("");
-      setIsSubmitting(false);
-    }
-  }, [isSubmitting, otp]);
-
-  useEffect(() => {
-    if (otp.length === 4) {
-      submitOTP();
-    }
-  }, [otp, submitOTP]);
-
-  const handleInputChange = (
-    index: number,
-    e: React.ChangeEvent<HTMLInputElement>
-  ) => {
-    const input = e.target;
-    const value = input.value;
-
-    // Only allow numbers from 0 to 9
-    const validValue = value.replace(/[^0-9]/g, "");
-
-    setOtp((prevOtp) => {
-      const newOtp =
-        prevOtp.substring(0, index) + validValue + prevOtp.substring(index + 1);
-      return newOtp;
-    });
-
-    if (
-      value.length === input.maxLength &&
-      index < inputRefs.current.length - 1
-    ) {
-      inputRefs.current[index + 1].focus();
-    }
-  };
-
-  const handleKeyDown = (
-    index: number,
-    e: React.KeyboardEvent<HTMLInputElement>
-  ) => {
-    const keyPressed = e.key;
-    const isNumberKey = /^[0-9]$/.test(keyPressed);
-
-    if (!isNumberKey && otp.charAt(index) === "") {
-      e.preventDefault();
-    }
-
-    if (keyPressed === "Backspace" && index > 0 && otp.charAt(index) === "") {
-      inputRefs.current[index - 1].focus();
+      console.log("resend otp");
+      // await sendOtp({ email: email });
+    } catch (errors) {
+      toast.error("somethingWentWrong");
     }
   };
 
   return (
-    <div className="flex flex-col items-center justify-center gap-4">
-      <div className="flex items-center justify-center">
-        <div className="flex space-x-2">
-          <input
-            ref={(el) => (inputRefs.current[0] = el!)}
-            type="text"
-            maxLength={1}
-            className="w-12 h-12 text-center border border-neutral-300 rounded-md disabled:animate-pulse"
-            value={otp.charAt(0) || ""}
-            disabled={isSubmitting}
-            onChange={(e) => handleInputChange(0, e)}
-            onKeyDown={(e) => handleKeyDown(0, e)}
-          />
-          <input
-            ref={(el) => (inputRefs.current[1] = el!)}
-            type="text"
-            maxLength={1}
-            className="w-12 h-12 text-center border border-neutral-300 rounded-md disabled:animate-pulse"
-            value={otp.charAt(1) || ""}
-            disabled={isSubmitting}
-            onChange={(e) => handleInputChange(1, e)}
-            onKeyDown={(e) => handleKeyDown(1, e)}
-          />
-          <input
-            ref={(el) => (inputRefs.current[2] = el!)}
-            type="text"
-            maxLength={1}
-            className="w-12 h-12 text-center border border-neutral-300 rounded-md disabled:animate-pulse"
-            value={otp.charAt(2) || ""}
-            disabled={isSubmitting}
-            onChange={(e) => handleInputChange(2, e)}
-            onKeyDown={(e) => handleKeyDown(2, e)}
-          />
-          <input
-            ref={(el) => (inputRefs.current[3] = el!)}
-            type="text"
-            maxLength={1}
-            className="w-12 h-12 text-center border border-neutral-300 rounded-md disabled:animate-pulse"
-            value={otp.charAt(3) || ""}
-            disabled={isSubmitting}
-            onChange={(e) => handleInputChange(3, e)}
-            onKeyDown={(e) => handleKeyDown(3, e)}
-          />
+    <div className="bg-[#14162E] m-auto min-h-screen flex flex-col items-center justify-center relative">
+      <motion.div
+        initial={{ opacity: 0 }}
+        animate={{ opacity: 1 }}
+        exit={{ opacity: 0 }}
+        transition={{ ease: "easeIn", duration: 0.25 }}
+        className="flex flex-col items-center justify-center h-screen gap-4"
+      >
+        <p className="text-3xl font-bold text-center">Email Verification</p>
+        <p className="text-lg font-medium text-center w-full">
+          <span className="text-neutral-500">
+            Enter below the 4 digits one-time password we sent to{" "}
+          </span>
+          <span className="text-secondary">{"test@gmail.com"}</span>
+          <span className="text-neutral-400 ">. </span>
+        </p>
+
+        <OtpInput email={"test@gmail.com"} />
+        <div className="flex flex-col items-center justify-center">
+          <div className="absolute z-10">
+            <button
+              className="flex items-center justify-center gap-2 text-neutral-500 hover:text-neutral-600"
+              onClick={handleResubmission}
+            >
+              Resend OTP
+            </button>
+          </div>
         </div>
-      </div>
+        <motion.div
+          initial={{ top: "21%", left: "6%", x: "0%", y: "0%", opacity: 1 }}
+          animate={{ top: "20%", left: "36%", x: "0%", y: "0%", opacity: 1 }}
+          transition={{ duration: 0.2, delay: 0 }}
+          className="absolute z-0"
+        >
+          <div className="w-60 h-60 bg-[#DDA82A] rounded-full blur-3xl opacity-50" />
+        </motion.div>
+        <motion.div
+          initial={{ top: "57%", left: "13%", x: "0%", y: "0%", opacity: 1 }}
+          animate={{ top: "57%", left: "47%", x: "0%", y: "0%", opacity: 1 }}
+          transition={{ duration: 0.2, delay: 0 }}
+          className="absolute z-0"
+        >
+          <div className="w-60 h-60 bg-[#4461F2] rounded-full blur-3xl opacity-50" />
+        </motion.div>
+      </motion.div>
     </div>
   );
 };
